REFACTOR: Moved messages and tests of feegrant module to own file.
