+++ pkg/indexer/decode/handle/feegrant.go
@@ -1,0 +1,17 @@
+package handle
+
+import (
+	"github.com/cosmos/cosmos-sdk/x/feegrant"
+	"github.com/dipdup-io/celestia-indexer/internal/storage"
+	storageTypes "github.com/dipdup-io/celestia-indexer/internal/storage/types"
+	"github.com/dipdup-io/celestia-indexer/pkg/types"
+)
+
+func MsgGrantAllowance(level types.Level, m *feegrant.MsgGrantAllowance) (storageTypes.MsgType, []storage.AddressWithType, error) {
+	msgType := storageTypes.MsgGrantAllowance
+	addresses, err := createAddresses(addressesData{
+		{t: storageTypes.MsgAddressTypeGranter, address: m.Granter},
+		{t: storageTypes.MsgAddressTypeGrantee, address: m.Grantee},
+	}, level)
+	return msgType, addresses, err
+}
