+++ pkg/indexer/decode/message_test.go
@@ -4,9 +4,7 @@ import (
 	"testing"
 
 	"cosmossdk.io/math"
-	cosmosCodecTypes "github.com/cosmos/cosmos-sdk/codec/types"
 	cosmosTypes "github.com/cosmos/cosmos-sdk/types"
-	cosmosFeegrant "github.com/cosmos/cosmos-sdk/x/feegrant"
 	cosmosSlashingTypes "github.com/cosmos/cosmos-sdk/x/slashing/types"
 	cosmosStakingTypes "github.com/cosmos/cosmos-sdk/x/staking/types"
 	"github.com/dipdup-io/celestia-indexer/internal/storage"
@@ -436,74 +434,6 @@ func TestDecodeMsg_SuccessOnMsgUnjail(t *testing.T) {
 		Namespace: nil,
 		Addresses: addressesExpected,
 	}
-	assert.NoError(t, err)
-	assert.Equal(t, uint64(0), dm.BlobsSize)
-	assert.Equal(t, msgExpected, dm.Msg)
-	assert.Equal(t, addressesExpected, dm.Addresses)
-}
-
-// MsgGrantAllowance
-
-func createMsgGrantAllowance() cosmosTypes.Msg {
-	m := cosmosFeegrant.MsgGrantAllowance{
-		Granter:   "celestia18r6ujzzkg6ku9sr39nxy4847q4qea5kg4a8pxv",
-		Grantee:   "celestia1vnflc6322f8z7cpl28r7un5dxhmjxghc20aydq",
-		Allowance: cosmosCodecTypes.UnsafePackAny(cosmosFeegrant.BasicAllowance{}),
-	}
-
-	return &m
-}
-
-func TestDecodeMsg_SuccessOnMsgGrantAllowance(t *testing.T) {
-	m := createMsgGrantAllowance()
-	blob, now := testsuite.EmptyBlock()
-	position := 4
-
-	dm, err := Message(m, blob.Height, blob.Block.Time, position, storageTypes.StatusSuccess)
-
-	addressesExpected := []storage.AddressWithType{
-		{
-			Type: storageTypes.MsgAddressTypeGranter,
-			Address: storage.Address{
-				Id:         0,
-				Height:     blob.Height,
-				LastHeight: blob.Height,
-				Address:    "celestia18r6ujzzkg6ku9sr39nxy4847q4qea5kg4a8pxv",
-				Hash:       []byte{0x38, 0xf5, 0xc9, 0x8, 0x56, 0x46, 0xad, 0xc2, 0xc0, 0x71, 0x2c, 0xcc, 0x4a, 0x9e, 0xbe, 0x5, 0x41, 0x9e, 0xd2, 0xc8},
-				Balance: storage.Balance{
-					Id:    0,
-					Total: decimal.Zero,
-				},
-			},
-		},
-		{
-			Type: storageTypes.MsgAddressTypeGrantee,
-			Address: storage.Address{
-				Id:         0,
-				Height:     blob.Height,
-				LastHeight: blob.Height,
-				Address:    "celestia1vnflc6322f8z7cpl28r7un5dxhmjxghc20aydq",
-				Hash:       []byte{0x64, 0xd3, 0xfc, 0x6a, 0x2a, 0x52, 0x4e, 0x2f, 0x60, 0x3f, 0x51, 0xc7, 0xee, 0x4e, 0x8d, 0x35, 0xf7, 0x23, 0x22, 0xf8},
-				Balance: storage.Balance{
-					Id:    0,
-					Total: decimal.Zero,
-				},
-			},
-		},
-	}
-
-	msgExpected := storage.Message{
-		Id:        0,
-		Height:    blob.Height,
-		Time:      now,
-		Position:  4,
-		Type:      storageTypes.MsgGrantAllowance,
-		TxId:      0,
-		Data:      structs.Map(m),
-		Namespace: nil,
-		Addresses: addressesExpected,
-	}
-
 	assert.NoError(t, err)
 	assert.Equal(t, uint64(0), dm.BlobsSize)
 	assert.Equal(t, msgExpected, dm.Msg)
